+++ tests/test_api.py
@@ -1,46 +1,69 @@
 """
-HappyBase API tests.
+HappyBase tests.
 """
 
-import os
 import collections
+import os
+import random
+import threading
 
-from nose.tools import (assert_dict_equal,
-                        assert_equal,
-                        assert_false,
-                        assert_in,
-                        assert_is_instance,
-                        assert_is_not_none,
-                        assert_not_in,
-                        assert_raises,
-                        assert_true)
+from nose.tools import (
+    assert_dict_equal,
+    assert_equal,
+    assert_false,
+    assert_in,
+    assert_is_instance,
+    assert_is_not_none,
+    assert_not_in,
+    assert_raises,
+    assert_true,
+)
 
-import happybase
+from happybase import Connection, ConnectionPool, NoConnectionsAvailable
 
 HAPPYBASE_HOST = os.environ.get('HAPPYBASE_HOST')
 HAPPYBASE_PORT = os.environ.get('HAPPYBASE_PORT')
 HAPPYBASE_COMPAT = os.environ.get('HAPPYBASE_COMPAT', '0.92')
+HAPPYBASE_TRANSPORT = os.environ.get('HAPPYBASE_TRANSPORT', 'buffered')
 KEEP_TABLE = ('HAPPYBASE_NO_CLEANUP' in os.environ)
 
 TABLE_PREFIX = 'happybase_tests_tmp'
 TEST_TABLE_NAME = 'test1'
 
+connection_kwargs = dict(
+    host=HAPPYBASE_HOST,
+    port=HAPPYBASE_PORT,
+    table_prefix=TABLE_PREFIX,
+    compat=HAPPYBASE_COMPAT,
+    transport=HAPPYBASE_TRANSPORT,
+)
+
+
+# Yuck, globals
 connection = table = None
+
+
+def maybe_delete_table():
+    if KEEP_TABLE:
+        return
+
+    if TEST_TABLE_NAME in connection.tables():
+        print "Test table already exists; removing it..."
+        connection.delete_table(TEST_TABLE_NAME, disable=True)
 
 
 def setup_module():
     global connection, table
-    connection = happybase.Connection(host=HAPPYBASE_HOST,
-                                      port=HAPPYBASE_PORT,
-                                      table_prefix=TABLE_PREFIX,
-                                      compat=HAPPYBASE_COMPAT)
+    connection = Connection(**connection_kwargs)
+
     assert_is_not_none(connection)
 
+    maybe_delete_table()
     cfs = {
         'cf1': {},
         'cf2': None,
         'cf3': {'max_versions': 1},
-        }
+    }
     connection.create_table(TEST_TABLE_NAME, families=cfs)
 
     table = connection.table(TEST_TABLE_NAME)
@@ -49,14 +72,19 @@ def setup_module():
 
 def teardown_module():
     if not KEEP_TABLE:
-        connection.disable_table(TEST_TABLE_NAME)
-        connection.delete_table(TEST_TABLE_NAME)
+        connection.delete_table(TEST_TABLE_NAME, disable=True)
     connection.close()
 
 
 def test_connection_compat():
     with assert_raises(ValueError):
-        happybase.Connection(compat='0.1.invalid.version')
+        Connection(compat='0.1.invalid.version')
+
+
+def test_timeout_arg():
+    Connection(
+        timeout=5000,
+        autoconnect=False)
 
 
 def test_enabling():
@@ -79,14 +107,14 @@ def test_prefix():
     assert_equal(connection.table('foobar').name, TABLE_PREFIX + '_foobar')
     assert_equal(connection.table('foobar', use_prefix=False).name, 'foobar')
 
-    c = happybase.Connection(autoconnect=False)
+    c = Connection(autoconnect=False)
     assert_equal('foo', c._table_name('foo'))
 
     with assert_raises(TypeError):
-        happybase.Connection(autoconnect=False, table_prefix=123)
+        Connection(autoconnect=False, table_prefix=123)
 
     with assert_raises(TypeError):
-        happybase.Connection(autoconnect=False, table_prefix_separator=2.1)
+        Connection(autoconnect=False, table_prefix_separator=2.1)
 
 
 def test_stringify():
@@ -128,6 +156,7 @@ def test_families():
 def test_put():
     table.put('r1', {'cf1:c1': 'v1', 'cf1:c2': 'v2', 'cf2:c3': 'v3'})
     table.put('r1', {'cf1:c4': 'v2'}, timestamp=2345678)
+    table.put('r1', {'cf1:c4': 'v2'}, timestamp=1369168852994L)
 
 
 def test_atomic_counters():
@@ -159,7 +188,7 @@ def test_batch():
 
     b = table.batch()
     b.put('row1', {'cf1:col1': 'value1',
-                  'cf1:col2': 'value2'})
+                   'cf1:col2': 'value2'})
     b.put('row2', {'cf1:col1': 'value1',
                    'cf1:col2': 'value2',
                    'cf1:col3': 'value3'})
@@ -391,6 +420,13 @@ def test_scan():
     scanner = table.scan(row_prefix='row', timestamp=123)
     assert_equal(0, calc_len(scanner))
 
+    scanner = table.scan(batch_size=20)
+    next(scanner)
+    next(scanner)
+    scanner.close()
+    with assert_raises(StopIteration):
+        next(scanner)
+
 
 def test_delete():
     row_key = 'row-test-delete'
@@ -417,3 +453,101 @@ def test_delete():
 
     table.delete(row_key)
     assert_dict_equal({}, table.row(row_key))
+
+
+def test_connection_pool_construction():
+    with assert_raises(TypeError):
+        ConnectionPool(size='abc')
+
+    with assert_raises(ValueError):
+        ConnectionPool(size=0)
+
+
+def test_connection_pool():
+
+    from thrift.transport.TTransport import TTransportException
+
+    def run():
+        name = threading.current_thread().name
+        print "Thread %s starting" % name
+
+        def inner_function():
+            # Nested connection requests must return the same connection
+            with pool.connection() as another_connection:
+                assert connection is another_connection
+
+                # Fake an exception once in a while
+                if random.random() < .25:
+                    print "Introducing random failure"
+                    connection.transport.close()
+                    raise TTransportException("Fake transport exception")
+
+        for i in xrange(50):
+            with pool.connection() as connection:
+                connection.tables()
+
+                try:
+                    inner_function()
+                except TTransportException:
+                    # This error should have been picked up by the
+                    # connection pool, and the connection should have
+                    # been replaced by a fresh one
+                    pass
+
+                connection.tables()
+
+        print "Thread %s done" % name
+
+    N_THREADS = 10
+
+    pool = ConnectionPool(size=3, **connection_kwargs)
+    threads = [threading.Thread(target=run) for i in xrange(N_THREADS)]
+
+    for t in threads:
+        t.start()
+
+    while threads:
+        for t in threads:
+            t.join(timeout=.1)
+
+        # filter out finished threads
+        threads = [t for t in threads if t.is_alive()]
+        print "%d threads still alive" % len(threads)
+
+
+def test_pool_exhaustion():
+    pool = ConnectionPool(size=1, **connection_kwargs)
+
+    def run():
+        with assert_raises(NoConnectionsAvailable):
+            with pool.connection(timeout=.1) as connection:
+                connection.tables()
+
+    with pool.connection():
+        # At this point the only connection is assigned to this thread,
+        # so another thread cannot obtain a connection at this point.
+
+        t = threading.Thread(target=run)
+        t.start()
+        t.join()
+
+
+if __name__ == '__main__':
+    import logging
+    import sys
+
+    # Dump stacktraces using 'kill -USR1', useful for debugging hanging
+    # programs and multi threading issues.
+    try:
+        import faulthandler
+    except ImportError:
+        pass
+    else:
+        import signal
+        faulthandler.register(signal.SIGUSR1)
+
+    logging.basicConfig(level=logging.DEBUG)
+
+    method_name = 'test_%s' % sys.argv[1]
+    method = globals()[method_name]
+    method()